Bug 1667202 - Allow scroll metadata creation even for APZ-disabled scrollframes. r=botond
APZ is not always enabled; for example, on the 1proc
chrome mochitests. However, WR still relies on getting a scroll metadata
for ASRs in this situation, so that it can inform WR of the scroll range
of scrollframes. This combination results in an assertion failure where
the scroll metadata is not generated but the caller expects it to be. This
patch fixes the problem by generating the scroll metadata even if APZ is
disabled. This should be harmless, and fixes the assertion
failure.

Differential Revision: https://phabricator.services.mozilla.com/D99671

UltraBlame original commit: 52a4a14e2d32071fe9b15cdfa7d6e400f0d913c5