Bug 1664444, make screenshots fission compatible and remove unused code r=sfoster
This patch removes the code that saves screenshots, as that is no longer supported.
We remove the drawing canvas work in shooter.js to make this fission compatible.
This work is repeated in the background, so we should instead rely on that to handle
downloading and copying a screenshot, as well as previewing full page and the visible
parts of the page.

Differential Revision: https://phabricator.services.mozilla.com/D98285

UltraBlame original commit: 90fd0f5a6bb226190fcc9312d68b3876f8c51c55