Bug 1637658 - P2. Create new decoder if initialization data doesn't match bytestream content. r=mattwoodrow,pehrsons
Many decoders are using the image size provided on initilization to determine how to create the decoded image. This can sometimes be incorrect like in WebRTC.

We now check if the image or display size is inconsistent with the bytestream data and if so force the creation of a new decoder to ensure it has valid initialization data.

This is what was already done for H264 as the initialization data was constructed from the SPS NAL (which is first required to instantiate a new decoder).

This is also a better fix for bug 1525230 which got around the bad aspect ratio. Following P1, there's no concept of aspect ratio left when the decoder is used outside of WebM container.

Differential Revision: https://phabricator.services.mozilla.com/D99602

UltraBlame original commit: 6180177f388a2f3aa7a36e516ec4d1faeb30db97