Bug 1680602 - Split `TrySkipAwait` into two opcodes r=arai
The TrySkipAwait opcode optimization is challenging to JIT compile because of
the complicate semantics of its return values. In order to make this
dramatically more simple to JIT implement, this patch splits TrySkipAwait into
two ops: `CanSkipAwait` and `MaybeExtractAwaitValue`.

The first op, `CanSkipAwait` does the preparation to ensure that it would be
legal to skip the await enqueue. The second op, `MaybeExtractAwaitValue` uses
the result of the `CanSkipAwait` to decide if it should leave the value alone
or instead replace it with the awaited value. This pair of ops is a bit easier
for Warp to model in MIR than a single op.

Differential Revision: https://phabricator.services.mozilla.com/D98756

UltraBlame original commit: db358fced289f38bb295d7ce3bd3f3b590702c7d