Bug 494175 part 3: Don't change the focusability of the full screen button in the video controls when fading out if the controls are in the tab order. r=Gijs
This was added to prevent visual confusion, since otherwise, the full screen button could be focused while the controls were hidden.
If the controls are in the tab order, they show when focused, so this is not necessary.

Differential Revision: https://phabricator.services.mozilla.com/D97530

UltraBlame original commit: 8a6d25038f813d0b8184a02e01b22b88940ba5c6