Bug 1681297 - Skip pre-write barriers triggered on background threads r=sfink
The problem here is that we can trigger this barrier when background
finalization destroys HeapPtrs to things in the atoms zone, since the atoms
zone may still be marking at this point. (Usualy cross-zone edges are stored in
the private pointer of a cross-compartment wrapper.)

To avoid the possiblity of races the patch checks the current thread when the
target thing is in the atoms zone. To make this work without pulling the whole
of Zone.h into Cell.h I had to move some of the zone's flags into shadow::Zone.
I'll tidy that a little in the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D99782

UltraBlame original commit: 5208d36713dbd97452656dd89f948654fec9e0fe