Bug 1682571 - Update prefs_file_was_invalid telemetry r=chutten,decoder
We are interested in knowing if and when the prefs file fails to read. Since Preferences silently handles this failure by resetting the prefs file (which is not ideal), the best way to find out if this is failing is via telemetry probe. The `prefs_file_was_invalid` telemetry already exists so it only needs to be updated.

Differential Revision: https://phabricator.services.mozilla.com/D99911

UltraBlame original commit: e08861664dc51c2d1b7e3fa8b43037050f49bc55