Bug 1683129 [wpt PR 26942] - Cannot balance columns based on overflow block-size., a=testonly
Automatic update from web-platform-tests
Cannot balance columns based on overflow block-size.

Replace TODO with comment.

If the inline-size of the contents of the multicol container is 0, an
empty overflow rectangle will be produced, even if block-size is
non-zero, and we'd fail balancing of blocks that are freely breakable
inside, if initial column block-size is 0. The space shortage reporting
machinery (which is what we use in subsequent column layout passes)
really only works for content that we cannot (or at least don't want to)
break inside. We'd also get log2(N)-ish relayouts (where N is the total
block-size of the contents).

So it looks like we need to keep CalculateColumnContentBlockSize(),
after all. I tried to get rid of it here, before I realized that it
would cause trouble:
https://chromium-review.googlesource.com/c/chromium/src/+/2597308

Add a test that demonstrates this. Note that
external/wpt/css/css-multicol/baseline-006.html also tests this, but
that's not the main point with that test.

Bug: 829028
Change-Id: I24516a5956f9e60f47c87cf51db2683feded1eeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2595540
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Cr-Commit-Position: refs/heads/master{#838116}

--

wpt-commits: f70ab3a9576f4ab463a523bde33e59a9f2914aea
wpt-pr: 26942

UltraBlame original commit: e353095e05412f0c532bd3abcbb097f4140d9654