Bug 1672862 - Part 3: Update TypedArray [[DefineProperty]]. r=tcampbell
Update TypedArray [[DefineProperty]] to match the current spec. This includes
throwing in `js::DefineTypedArrayElement` for out-of-bounds indices and
detached buffers. V8 already ships with this behaviour, so it should be
web-compatible for us to throw errors. We can't reuse the existing
`JSMSG_BAD_INDEX` error code, because that throws a `RangeError`.

As a side-effect of this change, `Object.seal()` will now throw non-empty
TypedArrays.

Differential Revision: https://phabricator.services.mozilla.com/D99382

UltraBlame original commit: b9b83bac0514965442ea64715bf1248db93068fa