Bug 1681985 - Add and call Fathom ruleset. r=zbraniecki
Remove calls to old heuristics, which are duplicated, improved, expanded, and balanced in the Fathom model. We'll come along in the next release and delete the dead code. We're leaving it in for now to keep it from bitrotting, in case we have to revert to it.

Differential Revision: https://phabricator.services.mozilla.com/D100141

UltraBlame original commit: 1ba9b39ff543cf15ed03f54a744c935695588431