Bug 1679735 require recent user gesture for getDisplayMedia() r=jib
Previous behavior was to allow requests if the user has ever interacted with
the window.  This patch changes from sticky activation to transient
activation.

WindowContext is used rather than Document for the
HasValidTransientUserGestureActivation() test because user activation is a
property of the Window/global.

The error reported when WindowContext is null is the same as that for missing
transient activation, which is keeping the existing behavior.
https://searchfox.org/mozilla-central/rev/66547980e8e8ca583473c74f207cae5bac1ed541/dom/base/Document.cpp#15806
i.e. a missing WindowContext is still treated as missing transient activation.

Differential Revision: https://phabricator.services.mozilla.com/D98275

UltraBlame original commit: 3faff9d75a5d431a481770916f1b7caae8e22187