Bug 1041377 - Part 3: Disable Backspace as keyboard shortcut for back/forward navigation on Windows and macOS. r=masayuki
This change aligns Firefox on Windows and macOS with the behavior of Firefox on Linux, Chrome, Edge, and Safari. The only mainstream browsers that navigate back on Backspace in 2020 are Firefox on Windows and macOS and IE11 on Windows.

Depends on D100532

Differential Revision: https://phabricator.services.mozilla.com/D100533

UltraBlame original commit: b1be428bb5d9fec2310c59d57b3051465fb874c0