Bug 1655052 - Simplify BytecodeEmitter::emitThisEnvironmentCallee. r=arai
The `emitThisEnvironmentCallee` helper is used when handling certain 'super'
operations. If the current scope is an eval or arrow, we instead use the
callee of the enclosing environment chain. When this is a case, the Parser
must already have marked the appropriate 'this' as closed-over and we can
simply assert that the scope has an environment.

Differential Revision: https://phabricator.services.mozilla.com/D100814

UltraBlame original commit: 374aebc92cc20b46180e0f7de8301874e48239a5