Bug 1685128 - Collect the nursery during idle time if appears to be underused r=sfink
The aim here is to detect when the nursery isn't being used much and collect it
so that it can shrink itself (this doesn't change the sizing heuristics). This
ensures the nursery doesn't hold on to memory that could be better used for
other things.

There are many possible approaches but I went for the simplest - a fixed
timeout. If the nursery is above minimum size and as not been collected for 2
seconds then we collect it (if we're idle). This sounds drastic but is actually
quite a long time in terms of nursery collections, and only happens during idle
time when there is no JS code running.

This appears to fix the AWSY regressions for the other patches that affected
the number of nursery collections.

Differential Revision: https://phabricator.services.mozilla.com/D100881

UltraBlame original commit: dc3237308ff0b09ae3471eb1f1bdf1cc93b596a4