Bug 1683329 [wpt PR 26953] - [LayoutNG] Make abspos' fragment encompass the height of its children, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Make abspos' fragment encompass the height of its children

In CL:2577757, we allowed fragmentation for abspos positioned using the
bottom property and with `height: auto`. There was one remaining issue
with the abspos fragment when a child of his had a break-rule - its
height was off, giving the impression that it didn't encompass all its
children. This CL resolves this issue.

The behavior of expanding the height to encompass all children will
cause the bottom property to be off from where it was originally
intended to be. That's a downside widely discussed offline between
Rossen (co-author of the CSS Fragmentation spec), Alison and I, but a
solution relatively acceptable to a problem that isn't NP-complete.
Gecko seems to do the same as well.

Bug: 1156250
Change-Id: Ie67f7aff9a64e7d813f346ad83cd6d4bcd5186e7

--

wpt-commits: 5cdc2d5c22e3af1bb66c1d952af900379e8568b9
wpt-pr: 26953

UltraBlame original commit: 197e8eebddb4cfcaacbc23f674df81c5193d30aa