Bug 1685599 - Don't rely on macosx-cross toolchain containing a mac native libLLVM. r=firefox-build-system-reviewers,andi,dmajor
The webrender wrench macos build, which is cross-compiled, needs a
macOS native libLLVM (a .dylib) to link against. The file is currently
part of the macosx-cross toolchain, but that was more incidental than
intentional. As we're going to change the macosx-cross toolchain in a
way that will remove the libLLVM.dylib, pull the file from the macOS
native clang.

Differential Revision: https://phabricator.services.mozilla.com/D101150

UltraBlame original commit: 3dcceef9ec6bed8f4aaa50fe423b9b02f7924ac6