Bug 1686021 - Use T::isTenured() rather than IsInsideNursery to simplify ShouldMark r=sfink
isTenured is implemented on both Cell and TenuredCell, so if T is not nursery
allocatable the check will be skipped.  This allows us to use a single template
function for ShouldMark.  I updated a couple of other uses of IsInsideNursery
too.

Depends on D101357

Differential Revision: https://phabricator.services.mozilla.com/D101358

UltraBlame original commit: bbc59e8488890bd42c54332abf329a7938ec1caa