Bug 1685758 - Increase stack size on win32 debug r=firefox-build-system-reviewers,mhentges
With clang 12, win32 debug tests run out of stack in the 1000-deep layout crashtests.

Usually when this type of thing happens, it's because the compiler made a drastically heavier inlining decision along the recursive part of the stack, with a difference in the hundreds of bytes, and we pay that price of that space 1000 times. In those cases we normally try to snip out the heavy piece with a MOZ_NEVER_INLINE.

However, in this case, the difference between old and new compilers was only something like 16 bytes of the cyclical piece. This suggests that we're already up pretty close against the limit, and we're only a bad day away from failing tests even with existing clang. This only happens in debug builds since they have lots of logging helpers on the stack. Release builds are still well under the limit.

So I propose to increase the stack limit in debug. While we normally have to be pretty careful about address space on win32, it should be less of a concern with these non-shipping builds.

Differential Revision: https://phabricator.services.mozilla.com/D101196

UltraBlame original commit: 68770df4c03a48d4c4a51bb83ecaf3d68a3d17e6