Bug 1686021 - Use T::isTenured() rather than IsInsideNursery to simplify ShouldMark r=sfink
isTenured is implemented on both Cell and TenuredCell, so if T is not nursery
allocatable the check will be skipped.  This allows us to use a single template
function for ShouldMark.  I updated a couple of other uses of IsInsideNursery
too.

Depends on D101357

Differential Revision: https://phabricator.services.mozilla.com/D101358

UltraBlame original commit: fa83b32c51b5e108beb40977d1939bdb9e455027