Bug 1599317 - [devtools] Fix performance issue in ObjectInspector when we have a lot of instances. r=bomsy.
At the moment, the ObjectInspector, while having its own reducer and actions, has its
state located in the state of its consumer application (e.g. webconsole, debugger).
This patch, by adding a `standalone` prop to the ObjectInspector, will allow the said
consumer to not have the object inspector state shared anymore.
This makes interaction in a given objectInspector much faster as only one instance
will have its lifecycle methods triggered. It also prevent cloning the consumer's state
on each object inspector interaction.
This new props is only used in the webconsole for now, as the debugger might need
to have access to its objectInspector state (this is also less of a problem, because
the debugger won't hold that much instances at a given point in time, unlike the console).

The DAMP data seems to indicate a 40% improvement in the console when expanding
an object while there are a thousand instances in the console.

Differential Revision: https://phabricator.services.mozilla.com/D99472

UltraBlame original commit: 8526cb232e9a3250fd3dd0103156033979b821f5