Bug 1686492 - Part 8: Update TypedArray.prototype.indexOf(). r=mgaudet
Updates the step references and adds an assertion which was added to the spec.

The updated spec wants us to perform a HasProperty check during the loop in
step 11.a. We can avoid this operation when we instead re-read the TypedArray
length after steps 5-6.

Depends on D101616

Differential Revision: https://phabricator.services.mozilla.com/D101617

UltraBlame original commit: dcfc69d91d0c62c3272cb32ff9998921f3f6dbd1