Bug 1682340 - [devtools] Prevent triggering Reader Mode with reverse search keyboard shortcut on Windows. r=bomsy.
We weren't calling preventDefault on the event handler, which was
allowing the browser to consume it, and toggle reader mode on windows,
which shares the same keyboard shortcut (F9).
A test is added to ensure we don't regress this (the test is failing without
the fix).

Differential Revision: https://phabricator.services.mozilla.com/D101919

UltraBlame original commit: 65d5af50131ebf540b1031c26bbf32cd7830f16e