Bug 1687394 - Simplify semantics of `requested_raster_space`. r=nical
Previously, it was possible to request a local/screen raster
space even if the owning stacking context didn't create an
offscreen surface.

This complicates various parts of the code, and also results
in a pass-through picture primitive being created (which we
want to remove as part of the work for #1684781).

With this change, it's only possible to change the raster space
when the enclosing stacking context creates an offscreen surface
for some other reason (e.g. 3d transform, filters etc).

Differential Revision: https://phabricator.services.mozilla.com/D102244

UltraBlame original commit: 1c7e09561958790302e409b906605c2d422f6fc6