Bug 1687672: Add BailoutKind::UnboxFolding r=jandem
This uses a similar approach to LICM / BoundsCheckHoisting / etc.

Note that if we would have hit the unbox anyway, then we'll hit it after bailing out and update the CacheIR, so this won't cause us to recompile additional times. This means that we could end up turning off FoldLoadsWithUnbox in cases where it's not the problem, but unlike LICM it's not making a big enough difference to justify special treatment.

Differential Revision: https://phabricator.services.mozilla.com/D102505

UltraBlame original commit: 2a97db6aa2eb7e186d39fc8e653fc8869d024837