Bug 1674431 - Size the non-native theme number spinner buttons more sensibly. r=spohl
Use the same width as the select arrow button makes sense to me and
fixes all the reported problems so far.

Also tweak the sizing / painting to take less height since these have a
background, so stuff like:

  <input type=number style="height: 100px">

Looks super ugly otherwise.

The new polygon coordinates are the result of scaling by 1.5 and
centering vertically the arrow, which gives the same visual effect for
the regular input without margin / padding, but prevents the issue
described above with too-tall inputs or inputs with lots of padding.

We might want to shrink them a bit more if needed, but I think this is
alright. For reference, Chrome uses 15px, but we remove the inline-end
padding by default so we compensate a bit that way.

If we don't like how this looks when the author specifies padding on the
input (see comment 13 for example), we might want to reconsider the
design so that the spinners look good with padding, but that's a
different task completely.

Differential Revision: https://phabricator.services.mozilla.com/D102754

UltraBlame original commit: a95d10a2de4dabc987709e67ca5d05be41441fb9