Bug 1685303: part 28) Add `virtual` `AccessibleCaretManager::MaybeFlushLayout`. r=TYLin,smaug
In the unit-tests, flushing is not done. This was prevented implicitly
by nulling `mPresShell`. This change makes that explicit by stubbing
`MaybeFlushLayout`.

Note that `IsTerminated`, which is called in the same context is already
virtual, so I expect this to not have a noticeable performance-impact.

Depends on D102607

Differential Revision: https://phabricator.services.mozilla.com/D102608

UltraBlame original commit: d361437d483b0d81a0b7e01b2be8c484d4ec6748