Bug 1675850 - Add MOZ_BACKGROUNDTASKS flag, {nsI}BackgroundTasks IDL and class. r=mossop,firefox-build-system-reviewers,mhentges
This adds a build flag (defaulting to on in Nightly builds) and an
`AppConstants.jsm` definition.

It lays the foundation for managing an active background task.  A
singleton, exposed as an XPCOM service, owns the task name and lives
for the life of the application.

This will be wired into the application startup path in future
commits.  It's separated to allow the changes to category registration
to come early in the commit sequence.

Differential Revision: https://phabricator.services.mozilla.com/D96481

UltraBlame original commit: a53b4ebb924f4e032320ae7e58706210b2076e37