Bug 1686687 - Convert touch_action_helpers.js to async/await style. r=botond
I removed the setup functions since we want to avoid having long-lasting
touchend listeners, and instead added promiseTouchEnd calls to the touch-
synthesizing functions (touchScrollRight and touchScrollDown). Also I
moved the add_completion_callback stuff out of the setup functions. In
the case where the test function was using touchActionSetup, the
completion callback is triggered by clicking on the "btnComplete" button
in tapComplete, so I moved the completion promise in there. In the
remaining cases, where the test function was using touchActionSetupAndWaitTestDone,
it's usually the final scroll of the test that triggers the completion
callback, and so I just inlined the promise into the start of the
test function.

Differential Revision: https://phabricator.services.mozilla.com/D102861

UltraBlame original commit: b07c3f33d04f35e850677b0715698888ae4bf00f