Bug 1686720 - [devtools] Add a test for :screenshot --selector . r=jdescottes.
This adds a test for the :screenshot command with the --selector argument.
We're also checking that using the command while selecting another target
in the context selector works as expected.
The test is skipped on non-fission since the context selector isn't enabled
when fission is disabled.

Differential Revision: https://phabricator.services.mozilla.com/D102597

UltraBlame original commit: 954440d77ac4698cadc5906af95ceac5495188b8