Bug 1688910 - Add an API for performing write barriers on structures that have GC thing children r=jandem
This adds an API for the case where we need to barrier all the edges out of a structure. This reduces direct usage of the barrier tracer and adds the assertions that are present in calls to the barrier functions.

Differential Revision: https://phabricator.services.mozilla.com/D103054

UltraBlame original commit: 1ff5310b9fa04d4eefa1adc067ccd7859381d016