Bug 1679440 - Add test ensuring crash reporter works for background tasks. r=mossop,gsvelto
This adds a test to ensure that crash reporting works inside of
`application --backgroundtask ...` invocations.  A test-only `crash`
task is added that uses `CrashTestUtils.jsm` to trigger a crash.  The
`xpcshell` harness invokes that background task and processes any
minidump and extras just as it does for existing `xpcshell`
subprocesses.

The test is homed in `toolkit/crashreporter`, rather than in
`toolkit/components/backgroundtasks`, because there is special
handling for `CrashTestUtils.jsm` and the `testcrasher` library.  It's
probably possible to make that infrastructure usable from multiple
locations but it seems low value.

Differential Revision: https://phabricator.services.mozilla.com/D98096

UltraBlame original commit: 073d44e79773b9c49ddbb241a08038d779ddb9e4