Bug 1686652 - Generalize the SIMD wormhole.  r=yury
Generalize the SIMD wormhole to both x86 and x64, and both baseline
and Ion, so that we can start experimenting with it on a broader
scale.

Change the config so that the feature will ride the trains,
preffed-off bydefault (the current setting).

Differential Revision: https://phabricator.services.mozilla.com/D101709

UltraBlame original commit: 58936f707b78c204f1dcff131691afd60485eed6