Bug 1689285 - Fix test_visibility.xhtml to not depend on theme values. r=eeejay
It's relying on the other menuitems being tall enough so that they
overflow 5em. When the non-native theme is enabled, this assumption
breaks because that theme doesn't add padding to XUL menuitems, and we
get the failure:

    wrong state bits for 'mi_file1.2.4' !got '0', expected 'offscreen'

Get what we want to test more explicitly by setting the height of the
menuitems appropriately.

Differential Revision: https://phabricator.services.mozilla.com/D103267

UltraBlame original commit: 3c73a3ef1672d51e10870c608601602b8d33abda