Bug 1689734 - Cleanup lazy-source in nsFrameMessageManager if using ScriptPreloader. r=kmag
For consistency with other uses of the ScriptPreloader, we ensure we use lazy
script source when compiling for the cache. This generates full bytecode for
the cache but also avoids saving a copy of the source. If the frame script
were to call Function.prototype.toString, the existing source-hook would load
it in the same way as the JSM loader. In practice, we avoid relying on this
in our chrome code. If we will not be writing to the cache, we can instead
compile with the JS syntax parser, similar to what we do for JSM loader.

Differential Revision: https://phabricator.services.mozilla.com/D103515

UltraBlame original commit: 4a4a2311a274ed96e56a1a9dc84b62eeb7cc5981