Bug 1689394 - Simplify TraceKind definition to reduce code generated by switch statements r=sfink
Currently we arrange for the low thre bits of the TraceKind value be set for all trace kinds greater than seven (see definition of JS::TraceKind). This is to save a branch creating a GCCellPtr in GCCellPtr::checkedCast.

This has the side effect of generating an 80 entry lookup table when we switch on TraceKind when there are only 12 distinct trace kinds. And clang doesn't use a branch in checkCast anyway.

The patch changes TraceKind so that the values increment by one each time. This affects performance by increasing inlining opportunities in the marking code.

Depends on D103501

Differential Revision: https://phabricator.services.mozilla.com/D103502

UltraBlame original commit: 7b413c436848849535ce7a4379798600a58c49df