Bug 1688713 - REX byte needs to consider all registers. r=nbp
This looks like a copy-paste problem: when I implemented the new
encoder for threeByteOp with Base+Index I copied code from an encoder
for Reg+offset, and thus my REX emitter considered only the base and
dest registers, not base, index, and dest.  In the case where only the
index register would require the REX prefix to be present we'd get the
wrong code.

Turns out it's not too hard to test this with one of the instructions
that uses the threeByteOp encoder.

Differential Revision: https://phabricator.services.mozilla.com/D103356

UltraBlame original commit: 8ebbe2c2368e9d96eb0d2bcaad1e5ebc89ca2e6e