Bug 1689698 - improve performance of DER.jsm r=mbirghan
This patch improves the performance of DER.jsm by changing readBytes to use
slice rather than repeatedly calling readByte.
Additionally, this patch removes the validation that the input to DERDecoder
consists of an array of integers in the range [0, 255]. This check is
unnecessary for all current consumers of DER.jsm because the input comes from
atob, which only outputs values in that range. If other consumers use DER.jsm
in the future, they will have to determine whether or not they need to validate
the input themselves first.

Differential Revision: https://phabricator.services.mozilla.com/D103838

UltraBlame original commit: eb3ee8b2a4d75a4ea321c82c95cc08d7bddf5322